// client/src/app/services/api/api.services.ts
export interface EntityDtoOfNullableGuid {
    id?: string | undefined;
}

export interface UserDto extends EntityDtoOfNullableGuid {
    firstname?: string | undefined;
    lastname?: string | undefined;
    profileImageUrl?: string | undefined;
    fullName?: string | undefined;
    email?: string | undefined;
    description?: string | undefined;
    isLocked: boolean;
    roleName?: string | undefined;
    createdBy?: string | undefined;
    updatedBy?: string | undefined;
    deletedBy?: string | undefined;
    createdAt?: string | undefined;
    updatedAt?: string | undefined;
    deletedAt?: string | undefined;
    invitedBy?: string | undefined;
}
// ...
export interface LoginResponseDto {
    token?: string | undefined;
    currentUser?: UserDto | undefined;
}

export interface LoginRequestDto {
    email: string;
    password: string;
}